# openapi.yaml
openapi: 3.0.3
info:
  title: Ledgerbook local server API
  version: "1.0.0"
  description: >
    Local API served by server.mjs / server.go for the Ledgerbook app.
    Binds to 127.0.0.1 only; there is no authentication — the protection is
    that the port is not reachable off-machine. Treat every endpoint as
    "trusted local caller".
servers:
  - url: http://127.0.0.1:8123
    description: Local Ledgerbook server (loopback only)
paths:
...
  /api/location:
    get:
      summary: Current ledger + snapshot locations and preset defaults
      responses:
        "200":
          description: Where data lives right now
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
    put:
      summary: Change the ledger file and/or snapshots folder (persisted to config.json)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ledgerPath: { type: string, description: Absolute path to the ledger .json file }
                snapshotDir: { type: string, description: Absolute path to the snapshots folder }
      responses:
        "200":
          description: New location applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  ledgerPath: { type: string }
                  snapshotDir: { type: string }
                  ledgerExists: { type: boolean }
        "400":
          description: Path was not absolute
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/ledgers:
    get:
      summary: List the ledger registry
      description: All ledgers the server knows, which one is current, the identity-colour palette, and which colours are taken.
      responses:
        "200":
          description: The registry
          content:
            application/json:
              schema:
                type: object
                properties:
                  ledgersRoot: { type: string, description: Folder where new ledgers are created }
                  current: { type: string, description: id of the current ledger }
                  palette:
                    type: array
                    items: { type: string, example: "#2f7d5b" }
                  usedColors:
                    type: array
                    items: { type: string }
                  ledgers:
                    type: array
                    items: { $ref: "#/components/schemas/LedgerEntry" }
    post:
      summary: Create a new ledger and switch to it
      description: Creates an empty ledger file under ledgersRoot, assigns an unused identity colour (or the one supplied, if free), and makes it current.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "Business" }
                color: { type: string, description: Optional; must be unused, else 409 }
                dir: { type: string, description: Optional absolute folder to create the ledger in; defaults to ledgersRoot }
      responses:
        "200":
          description: Created and switched
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, id: { type: string }, current: { type: string } }
        "409":
          description: Requested colour already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      summary: Mutate a ledger — switch, rename, recolour, or remove
      description: >
        op=switch makes id current; op=rename sets its name; op=color sets its identity colour
        (409 if that colour is used by another ledger); op=remove unregisters it (the file on disk is
        left untouched) and falls back to another ledger, creating a default one if none remain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [op, id]
              properties:
                op: { type: string, enum: [switch, rename, color, remove] }
                id: { type: string }
                name: { type: string }
                color: { type: string }
                snapshots: { type: boolean, description: "op=remove only: also delete this ledger's snapshots (the file itself always stays)" }
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, current: { type: string } }
        "404": { description: No such ledger }
        "409":
          description: Colour already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/book:
    get:
      summary: Read the whole ledger
      description: >
        Returns the ledger JSON, or the literal null if no file exists yet. The ETag response
        header is the file's current revision — the client keeps it and sends it back as If-Match
        on the next write (optimistic concurrency).
      responses:
        "200":
          description: The ledger (or null)
          headers:
            ETag:
              description: Opaque revision token of the returned file (send back as If-Match on PUT).
              schema: { type: string, example: '"9f2a1c4b8e0d7a63"' }
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/Ledger" }
                  - { type: "null" }
    put:
      summary: Write the whole ledger
      description: >
        Overwrites the ledger file and writes a deduped snapshot to the snapshots folder.
        If an If-Match header is sent, the file on disk must still be at that revision, otherwise
        the write is refused with 409 and the file is left untouched — the fresher version wins.
        Omitting If-Match forces an unconditional overwrite.
      parameters:
        - in: header
          name: If-Match
          required: false
          description: Revision the client last read (the ETag from GET). Omit to overwrite unconditionally.
          schema: { type: string, example: '"9f2a1c4b8e0d7a63"' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Ledger" }
      responses:
        "204":
          description: Saved (no content)
          headers:
            ETag:
              description: New revision token of the just-written file.
              schema: { type: string, example: '"1b7e3d9a2c4f6058"' }
        "409":
          description: Conflict — the file on disk is newer than the If-Match revision; not written.
          headers:
            ETag:
              description: Current revision on disk (what the client should reconcile against).
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { description: Write error }
  /api/rates:
    get:
      summary: Exchange-rate lookup (proxied)
      description: Uses TWELVEDATA_API_KEY if set, else a free fiat source. Server-side call keeps the key off the browser.
      parameters:
        - in: query
          name: symbol
          required: true
          schema: { type: string, example: "EUR/USD" }
      responses:
        "200":
          description: Rate for the symbol
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbol: { type: string, example: "EUR/USD" }
                  rate: { type: number, example: 1.0837 }
        "502": { description: Upstream error }
  /api/pick:
    get:
      summary: Native macOS file/folder chooser (via osascript)
      description: Opens a native dialog on the machine running the server. Returns the chosen POSIX path, or cancelled=true.
      parameters:
        - in: query
          name: kind
          schema: { type: string, enum: [file, folder], default: folder }
      responses:
        "200":
          description: Chosen path or cancellation
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties: { path: { type: string, example: "/Users/you/Desktop/ledger.json" } }
                  - type: object
                    properties: { cancelled: { type: boolean, example: true } }
...
components:
  schemas:
    Location:
      type: object
      properties:
        ledgerPath: { type: string }
        snapshotDir: { type: string }
        ledgerExists: { type: boolean }
        defaults:
          type: object
          properties:
            icloud: { type: string }
            home: { type: string }
            cwd: { type: string }
    LedgerEntry:
      type: object
      description: One ledger in the registry (the registry lives in config.json, not in a ledger file).
      properties:
        id: { type: string }
        name: { type: string, example: "Personal" }
        path: { type: string }
        color: { type: string, example: "#2f7d5b", description: Unique identity colour }
        exists: { type: boolean }
        current: { type: boolean }
        lastTx: { type: string, nullable: true, description: Date of the latest transaction in this ledger (YYYY-MM-DD), or null }
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
    Ledger:
      type: object
      description: The whole double-entry book. accountId values on lines reference accounts[].id.
      required: [types, accounts, transactions, settings]
      properties:
        schemaVersion: { type: integer, example: 1 }
        types:
          type: array
          items:
            type: object
            required: [id, name, class]
            properties:
              id: { type: string }
              name: { type: string }
              class: { type: string, enum: [asset, liability, equity] }
              comment: { type: string }
        accounts:
          type: array
          items:
            type: object
            required: [id, name, typeId, currency]
            properties:
              id: { type: string }
              name: { type: string }
              typeId: { type: string }
              currency: { type: string, example: "USD" }
              monetary: { type: boolean, description: Cash-like (revalued by FX) vs carried at historical cost }
              comment: { type: string }
              deleted: { type: boolean, description: Soft-deleted (archived after a currency conversion) }
        transactions:
          type: array
          items:
            type: object
            required: [id, date, description, lines]
            properties:
              id: { type: string }
              date: { type: string, format: date }
              description: { type: string }
              comment: { type: string }
              vehicle: { type: string, example: "USD", description: Pivot currency the frozen leg rates are expressed in }
              lines:
                type: array
                items:
                  type: object
                  required: [accountId, debit, credit]
                  properties:
                    accountId: { type: string }
                    debit: { type: number }
                    credit: { type: number }
                    rate: { type: number, description: Frozen value of 1 unit of the leg's currency in the vehicle, at save time }
        settings:
          type: object
          properties:
            currency: { type: string, example: "USD", description: Display/base currency }
            vehicle: { type: string, example: "USD", description: Pivot currency for cross-conversions }
            rates:
              type: object
              additionalProperties: { type: number }
            ratesUpdated: { type: string }
            theme: { type: string, enum: [auto, light, dark] }
            font: { type: string, description: Per-ledger font choice }